Mod24/src/br/com/rpires/service/ContratoService.java
package br.com.rpires.service;

import br.com.rpires.dao.IContratoDao;

/**
 * @author rodrigo.pires
 */
public class ContratoService implements IContratoService {

    private IContratoDao contratoDao;

    public ContratoService(IContratoDao dao) {
        this.contratoDao = dao;
    }

    public ContratoService() {

    }

    @Override
    public String salvar() {
        contratoDao.salvar();
        return "Sucesso";
    }

    @Override
    public String buscar() {
        return "Sucesso";
    }
    public String excluir(){
        return "Sucesso";
    }
    public String atualizar(){
        return "Sucesso";
    }
}
